+++ fiddles/angular2-cli/fiddle-0029-EntAgGridChartSeriesClick/src/app/bar-chart/bar-chart.component.ts
@@ -6,22 +6,29 @@ import {
   KeyValueDiffers,
   DoCheck,
   ElementRef,
-  OnInit
+  OnInit,
+  Output,
+  EventEmitter
 } from '@angular/core';
+
+import * as _ from 'lodash';
 
 import {
   ChartRangeParams,
   ColDef,
   RowNode,
   ChartType,
-  CellRangeParams
+  CellRangeParams,
+  BarChartOptions as AgBarChartOptions
 } from 'ag-grid-community';
 
 
 import {
+  BarChartEvent,
+  BarChartEventTypeEnum,
   BarChartOptionKeysEnum,
   BarChartOptions
-} from './bar-chart-options';
+} from './bar-chart';
 
 
 import "ag-grid-enterprise/chartsModule";
@@ -35,6 +42,7 @@ import "ag-grid-enterprise/chartsModule";
 export class BarChartComponent implements OnChanges, DoCheck, OnInit {
 
   @Input() options: BarChartOptions;
+  @Output() events: EventEmitter<BarChartEvent>;
 
   //#region Internal Properties
 
@@ -43,10 +51,12 @@ export class BarChartComponent implements OnChanges, DoCheck, OnInit {
   private _differ: KeyValueDiffer<any, any>;
   private _columnDefs: ColDef[] = [];
   private _rowData: RowNode[] = [];
-  private _id;
-
+  private _id: string;
+  private _chartComponent: any;
+  private _chartOptions: AgBarChartOptions;
   private _onCreateChartContainer: Function;
-
+  private _onProcessChartOptions: Function;
+  private _chartHeight: number;
 
   private get _fields(): string[] {
     let values: string[] = [];
@@ -90,6 +100,8 @@ export class BarChartComponent implements OnChanges, DoCheck, OnInit {
 
   constructor(private _differs: KeyValueDiffers, private _elementRef : ElementRef) {
     this._onCreateChartContainer = (chartRef: any) => this.onCreateChartContainer(chartRef);
+    this._onProcessChartOptions = (params: any) => this.onProcessChartOptions(params);
+    this.events = new EventEmitter();
   }
 
   //#endregion
@@ -128,7 +140,8 @@ export class BarChartComponent implements OnChanges, DoCheck, OnInit {
   //#region OnInit Implementation
 
   ngOnInit(): void {
-    this._containerEl = this._elementRef.nativeElement.querySelector('.chart-container');
+    this._containerEl = this._elementRef.nativeElement.querySelector('.chart-wrapper-body');
+    this._chartHeight = Math.floor(.90 * window.innerHeight);
   }
 
   //#endregion
@@ -139,13 +152,49 @@ export class BarChartComponent implements OnChanges, DoCheck, OnInit {
     if (this._id && this._isPopulated && this._containerEl) {
       setTimeout(() => {
         params.api.chartRange(this._chartRangeParams);
+        this.events.emit({
+          type: BarChartEventTypeEnum.ready,
+          data: null
+        });
       }, 500);
     }
   }
 
   onProcessChartOptions(params:any): any {
-    var options = params.options;
+    if (this._chartOptions) {
+      return _.merge({}, params.options, this._chartOptions);
+    } else {
+      return this._defaultChartOptions(params.options);
+    }
+  }
 
+
+  onCreateChartContainer(chartRef) {
+    if (this._containerEl) {
+      this._chartComponent = chartRef.chartElement.__agComponent;
+      chartRef.chartElement.addEventListener('click', ($event: any) => this.onChartElementClick($event));
+      this._containerEl.appendChild(chartRef.chartElement);
+    }
+  }
+
+  onChartElementClick(event: any) {
+    if (this._chartComponent) {
+      const seriesNode = this._chartComponent.chartProxy.chart.pickSeriesNode(event.offsetX, event.offsetY);
+      if (seriesNode) {
+        this.events.emit({
+          type: BarChartEventTypeEnum.seriesNodeClick,
+          data: seriesNode.node.datum
+        });
+      }
+    }
+  }
+
+
+  //#endregion
+
+  //#region Internal Methods
+
+  private _defaultChartOptions(options: any): any {
     options.height = 500;
     options.width = 1000;
     options.title = {
@@ -173,9 +222,8 @@ export class BarChartComponent implements OnChanges, DoCheck, OnInit {
     options.tooltipClass = "my-tool-tip-class";
     options.legendPosition = "bottom";
     options.legendPadding = 20;
-    options.legend = {
-      enabled: false
-    };
+    options.legend.enabled = false;
+
     var xAxis = options.xAxis;
     xAxis.lineWidth = 2;
     xAxis.lineColor = "gray";
@@ -245,27 +293,18 @@ export class BarChartComponent implements OnChanges, DoCheck, OnInit {
       blur: 8
     };
     seriesDefaults.tooltipRenderer = function(params) {
+      console.log('tooltipRenderer', params);
+
       var xField = params.xField;
       var yField = params.yField;
       var x = params.datum[xField];
       var y = params.datum[yField];
       return "<b>" + xField.toUpperCase() + ":</b> " + x + "<br/><b>" + yField.toUpperCase() + ":</b> " + y;
     };
+
     return options;
   }
 
-
-  onCreateChartContainer(chartRef) {
-    console.log(chartRef);
-    if (this._containerEl) {
-      this._containerEl.appendChild(chartRef.chartElement);
-    }
-  }
-
-
-  //#endregion
-
-  //#region Internal Methods
 
   private _applyChange(item: any): void {
     switch (item.key) {
@@ -277,6 +316,9 @@ export class BarChartComponent implements OnChanges, DoCheck, OnInit {
         break;
       case BarChartOptionKeysEnum.rows:
         this._rowData = this.options.rows;
+        break;
+      case BarChartOptionKeysEnum.chartOptions:
+        this._chartOptions = this.options.chartOptions;
         break;
     }
   }